fix:style,feat:calculator
